STRCONCAT narrows its stack probe on purpose
Every use of STRCONCAT printed C4267 at /W3: the macro hands a size_t to check_stack_overflow,
which takes u32 and is exported from xrCore by ordinal under that decorated name, so the
parameter stays what it is and the macro casts. The value is a stack buffer length, far below
the limit either way. parse_script_namespace copied a prefix with strncpy and terminated it by
hand, which is memcpy and no C4996. script_engine.cpp builds clean with that, and so does every
other file whose only warning was the macro.

+++ src/xrCore/string_concatenations.h
@@ -72,7 +72,7 @@ pstr strconcat(char (&dest)[Size], const Args... sources)
     {                                                                       \
         xray::core::detail::string_tupples tupplesUniqueId(__VA_ARGS__);    \
         size_t bufferSize = tupplesUniqueId.size();                         \
-        xray::core::detail::check_stack_overflow(bufferSize);               \
+        xray::core::detail::check_stack_overflow(static_cast<u32>(bufferSize)); \
         (dest) = static_cast<pstr>(xr_alloca(bufferSize));                  \
         tupplesUniqueId.concat(dest);                                       \
     } while (0)
